refactor: move role and workflow selector
