+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/AppPermissions/types.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { ServerError, ACCESS_TYPE_MAP } from '@devtron-labs/devtron-fe-common-lib'
+import { ServerError, ACCESS_TYPE_MAP, SelectPickerOptionType } from '@devtron-labs/devtron-fe-common-lib'
 import { JobList } from '../../../../../../components/Jobs/Types'
 import { DirectPermissionsRoleFilter } from '../../../types'
 
@@ -55,4 +55,9 @@ export interface DirectPermissionRowProps
     handleDirectPermissionChange: (...rest) => void
     index: number
     removeRow: (index: number) => void
+}
+
+export interface WorkflowListType {
+    options: SelectPickerOptionType[]
+    loading: boolean
 }
